corrected update and delete user tests

+++ src/handlers/deleteUser.js
@@ -5,8 +5,14 @@ module.exports = async (req, h) => {
   const userId = req.params.id;
 
   try {
-    const deletedAt = new Date();
-    await db.deleteUser([deletedAt, userId]);
+    const users = await db.findById(userId);
+
+    if (users.length > 0) {
+      const deletedAt = new Date();
+      await db.deleteUser([deletedAt, userId]);
+    } else {
+      return new Boom.conflict('Unknown id');
+    }
   } catch (err) {
     req.log(['delete', 'user', 'error'], err);
     return new Boom.badImplementation(err);
